+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Uliweb Change Log
   all blocks defined in template. With `--with-filename` will display template filename
   also.
 * ini replace variable will be moved to the end of section
+* Add content-range support for filedown thanks to zhangchunlin
 
 0.2.1 Version
 -----------------